Make hideWholePostExcerpts an optional feature
When I used this module, I found that all my short blog posts were showing only the title and the date on the homepage.

I tracked it down to the change introduced in this pull request: https://github.com/chekun/hexo-excerpt/pull/6

The change seems very odd to me, and definitely isn't what I want from this module so I added a configuration option to switch between the behaviour introduced in that pull request and the previous behaviour.

Initially I set the default config value to behave post-#6 but the tests were failing. Switching to use the older behaviour by default caused the tests to pass again, so I made that the default.